fix: error message if file exists

+++ src/fs/create.js
@@ -7,9 +7,13 @@ const __dirname = path.dirname(__filename);
 const pathToFileToCreate = path.join(__dirname, 'files', 'fresh.txt');
 
 const create = async () => {
+  if (fs.existsSync(pathToFileToCreate)) {
+    throw new Error('FS operation failed');
+  }
+
   fs.writeFile(pathToFileToCreate, 'I am fresh and young', (error, data) => {
-    if (error) throw 'FS operation failed';
-    console.log(data);
+    if (error) throw new Error('FS operation failed');
+    console.log('File "fresh.txt" created');
   });
 };
